Update and rename test_securite.py to test_branches.py
